dotCMS/core#23011 Move the dotcms library to new NX system (#23122)
* Move lib to new nx tsc lib

* Update readme

* Fix versioning

* Bring back testing changes

* Bring back typedoc

* Fix path

* Move dotcms-models to new lib nx

* Update version
