+++ core-web/libs/dotcms-models/.eslintrc.json
@@ -3,33 +3,15 @@
     "ignorePatterns": ["!**/*"],
     "overrides": [
         {
-            "files": ["*.ts"],
-            "extends": [
-                "plugin:@nrwl/nx/angular",
-                "plugin:@angular-eslint/template/process-inline-templates"
-            ],
-            "rules": {
-                "@angular-eslint/directive-selector": [
-                    "error",
-                    {
-                        "type": "attribute",
-                        "prefix": "dotcms",
-                        "style": "camelCase"
-                    }
-                ],
-                "@angular-eslint/component-selector": [
-                    "error",
-                    {
-                        "type": "element",
-                        "prefix": "dotcms",
-                        "style": "kebab-case"
-                    }
-                ]
-            }
+            "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
+            "rules": {}
         },
         {
-            "files": ["*.html"],
-            "extends": ["plugin:@nrwl/nx/angular-template"],
+            "files": ["*.ts", "*.tsx"],
+            "rules": {}
+        },
+        {
+            "files": ["*.js", "*.jsx"],
             "rules": {}
         }
     ]
