+++ core-web/libs/dotcms-models/README.md
@@ -1,5 +1,11 @@
 # dotcms-models
 
-Collection of DotCMS interfaces and types to use internally in our app and libs.
-
 This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build dotcms-models` to build the library.
+
+## Running unit tests
+
+Run `nx test dotcms-models` to execute the unit tests via [Jest](https://jestjs.io).
