+++ core-web/libs/dotcms-models/jest.config.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
 export default {
-    displayName: 'dotcms',
+    displayName: 'dotcms-models',
     preset: '../../jest.preset.js',
     globals: {
         'ts-jest': {
@@ -11,5 +11,5 @@ export default {
         '^.+\\.[tj]s$': 'ts-jest'
     },
     moduleFileExtensions: ['ts', 'js', 'html'],
-    coverageDirectory: '../../coverage/libs/dotcms'
+    coverageDirectory: '../../coverage/libs/dotcms-models'
 };
