+++ core-web/libs/dotcms-models/package.json
@@ -1,4 +1,5 @@
 {
     "name": "@dotcms/dotcms-models",
-    "version": "0.0.12"
+    "version": "0.0.12",
+    "type": "commonjs"
 }
