+++ core-web/libs/dotcms-models/project.json
@@ -1,30 +1,32 @@
 {
     "$schema": "../../node_modules/nx/schemas/project-schema.json",
-    "projectType": "library",
     "sourceRoot": "libs/dotcms-models/src",
-    "prefix": "dotcms",
+    "projectType": "library",
     "targets": {
         "build": {
-            "executor": "@nrwl/angular:ng-packagr-lite",
+            "executor": "@nrwl/js:tsc",
+            "outputs": ["{options.outputPath}"],
             "options": {
+                "outputPath": "dist/libs/dotcms-models",
+                "main": "libs/dotcms-models/src/index.ts",
                 "tsConfig": "libs/dotcms-models/tsconfig.lib.json",
-                "project": "libs/dotcms-models/ng-package.json"
-            },
-            "configurations": {
-                "production": {
-                    "tsConfig": "libs/dotcms-models/tsconfig.lib.prod.json"
-                }
+                "assets": ["libs/dotcms-models/*.md"]
             }
         },
         "lint": {
             "executor": "@nrwl/linter:eslint",
+            "outputs": ["{options.outputFile}"],
             "options": {
-                "lintFilePatterns": [
-                    "libs/dotcms-models/src/**/*.ts",
-                    "libs/dotcms-models/src/**/*.html"
-                ]
-            },
-            "outputs": ["{options.outputFile}"]
+                "lintFilePatterns": ["libs/dotcms-models/**/*.ts"]
+            }
+        },
+        "test": {
+            "executor": "@nrwl/jest:jest",
+            "outputs": ["coverage/libs/dotcms-models"],
+            "options": {
+                "jestConfig": "libs/dotcms-models/jest.config.ts",
+                "passWithNoTests": true
+            }
         }
     },
     "tags": []
