+++ core-web/libs/dotcms-models/tsconfig.lib.json
@@ -2,18 +2,9 @@
     "extends": "./tsconfig.json",
     "compilerOptions": {
         "outDir": "../../dist/out-tsc",
-        "target": "es2015",
         "declaration": true,
-        "declarationMap": true,
-        "inlineSources": true,
-        "types": [],
-        "lib": ["dom", "es2018"]
+        "types": []
     },
-    "angularCompilerOptions": {
-        "skipTemplateCodegen": true,
-        "strictMetadataEmit": true,
-        "enableResourceInlining": true
-    },
-    "exclude": [],
-    "include": ["**/*.ts"]
+    "include": ["**/*.ts"],
+    "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
 }
