--- core-web/libs/dotcms-models/typedoc.json
@@ -1,10 +1,0 @@
-{
-    "name": "DotCMS Models",
-    "out": "../../dist-docs/docs/dotcms-models",
-    "entryPoints": ["./src/index.ts"],
-    "readme": "./libs/dotcms/README.md",
-    "excludePrivate": true,
-    "tsconfig": "./tsconfig.lib.json",
-    "excludeExternals": true,
-    "exclude": ["**/*.spec.ts"]
-}
