+++ core-web/libs/dotcms/package.json
@@ -1,4 +1,5 @@
 {
     "name": "dotcms",
-    "version": "0.0.19"
+    "version": "0.0.19",
+    "type": "commonjs"
 }
