+++ core-web/libs/dotcms/src/lib/api/DotApiForm.ts
@@ -2,7 +2,7 @@ import { DotCMSFormConfig } from '../models';
 import { DotApiContentType } from './DotApiContentType';
 
 // This import allow us to use the type for the form: HTMLDotFormElement
-import { DotCMSContentType } from '@dotcms/dotcms-models';
+import { DotCMSContentType } from 'dotcms-models';
 
 /**
  * Creates and provide methods to render a DotCMS Form
